refactor: shared/components 하위 index.tsx 파일명을 컴포넌트명으로 통일
