+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './index.css';
 import { CartPage } from './features/Cart/pages/CartPage';
-import { AppLayout } from './shared/components/AppLayout';
+import { AppLayout } from './shared/components/AppLayout/AppLayout';
 import { ToastProvider } from './shared/context/ToastProvider';
 
 export const App = () => {
