+++ src/shared/components/CheckBox/CheckBox.tsx
